requirements.txt
noise==1.2.2
numpy==1.23.1
pygame==2.0.1
Pillow==8.3.1